+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * 4.16.0
 
   * New Features:
-    * **global**: option `output` to redirect to a file instead of `stdout`
+    * **global**: option `output` to redirect result to a file instead of `stdout`
     * **global**: new option `silent_insecure`
     * `config`: keys added to `config ascp info`
     * `config`: added command `pubkey` to extract public key from private key
@@ -20,6 +20,7 @@
   * Breaking Changes:
     * `shares`: option `type` for users and groups is replaced with mandatory positional argument with same value.
     * `aoc`, `faspex`: package `recv` command changed to `receive`, for consistency with faspex5 (`recv` is now an alias command)
+    * `direct` agent: better support for WSS
 
 * 4.15.0